Move Deck<*>Item implementation into Deck.cpp
Due to their template base, separate source files are trivial and rather
unecessary, so the type-specific implementations are moved into
Deck.cpp.
